Split field classes into separate files
