+++ Fields/EmptyDbOptions.cs
@@ -1,0 +1,5 @@
+namespace Keystone4NET.Fields;
+
+public class KeystoneEmptyDbOptions : KeystoneFieldDbOptions
+{
+}
